# Conflicts:
#	afirma-simple/src/main/java/es/gob/afirma/standalone/protocol/ProtocolInvocationLauncherSign.java
#	afirma-simple/src/main/java/es/gob/afirma/standalone/protocol/ProtocolInvocationLauncherSignAndSave.java

+++ afirma-keystores-mozilla/src/main/java/es/gob/afirma/keystores/mozilla/apple/ShellScript.java
@@ -122,7 +122,7 @@ public class ShellScript {
 			output = AOUtil.getDataFromInputStream(standardStream);
 		}
 
-		if (this.deleteFile) {
+		if (this.deleteFile && this.scriptFile != null) {
 			Files.delete(this.scriptFile.toPath());
 		}
